aoc: Refactor day01 solution to use a recursive function for part 2
The solution to part 2 was previously implemented using a `list.fold` which was inefficient and difficult to read.  This commit refactors the solution to use a recursive function `travel` which makes the logic clearer and more efficient.  The `next_pos` function was also extracted to improve readability.  The input is now hardcoded for testing purposes.